refactor: Refactor entity classed
