+++ src/main/java/com/wanted/preonboarding/applicant/entity/Applicant.java
@@ -1,6 +1,7 @@
-package com.wanted.preonboarding.applicant;
+package com.wanted.preonboarding.applicant.entity;
 
 import jakarta.persistence.*;
+import org.hibernate.annotations.ColumnDefault;
 
 import java.util.HashSet;
 
@@ -11,8 +12,13 @@ public class Applicant {
     @Column(name = "applicant_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String name;
+
     @OneToMany(mappedBy = "applicant")
     private HashSet<Application> applications = new HashSet<>();
 
+    @Column(name = "is_deleted")
+    @ColumnDefault("'N'")
+    private String isDeleted;
 }
