+++ src/main/java/com/wanted/preonboarding/applicant/entity/Application.java
@@ -1,7 +1,8 @@
-package com.wanted.preonboarding.applicant;
+package com.wanted.preonboarding.applicant.entity;
 
-import com.wanted.preonboarding.recruitment.Recruitment;
+import com.wanted.preonboarding.recruitment.entity.Recruitment;
 import jakarta.persistence.*;
+import org.hibernate.annotations.ColumnDefault;
 
 @Entity
 public class Application {
@@ -10,11 +11,16 @@ public class Application {
     @Column(name = "application_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @ManyToOne
+
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "applicant_id")
     private Applicant applicant;
 
-    @ManyToOne
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "recruitment_id")
     private Recruitment recruitment;
+
+    @Column(name = "is_deleted")
+    @ColumnDefault("'N'")
+    private String isDeleted;
 }
