+++ src/main/java/com/wanted/preonboarding/recruitment/entity/Company.java
@@ -1,6 +1,7 @@
-package com.wanted.preonboarding.recruitment;
+package com.wanted.preonboarding.recruitment.entity;
 
 import jakarta.persistence.*;
+import org.hibernate.annotations.ColumnDefault;
 
 @Entity
 public class Company {
@@ -9,9 +10,15 @@ public class Company {
     @Column(name = "company_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
+
     private String name;
+
     private String country;
+
     private String region;
+
+    @Column(name = "is_deleted")
+    @ColumnDefault("'N'")
     private String isDeleted;
 
 }
